requirements/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.2
billiard==3.6.4.0
black==20.8b1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.2
django-model-utils==4.1.1
django-redis==4.12.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
kombu==5.0.2
mypy-extensions==0.4.3
pathspec==0.8.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.5
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
